@@ -2,40 +2,34 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
+
     <parent>
-        <groupId>com.blackphoenixproductions</groupId>
-        <artifactId>whiteboard</artifactId>
-        <version>1.6.0</version>
+        <groupId>org.springframework.boot</groupId>
+        <artifactId>spring-boot-starter-parent</artifactId>
+        <version>2.6.5</version>
     </parent>
 
+    <groupId>com.blackphoenixproductions</groupId>
     <artifactId>forum-backend</artifactId>
-    <version>1.6.0</version>
+    <version>2.0.0</version>
     <name>forum-backend</name>
     <description>backend api forum</description>
 
     <properties>
-        <java.version>1.8</java.version>
-        <maven.compiler.source>8</maven.compiler.source>
-        <maven.compiler.target>8</maven.compiler.target>
+        <java.version>11</java.version>
+        <maven.compiler.source>11</maven.compiler.source>
+        <maven.compiler.target>11</maven.compiler.target>
         <version.spring-boot-starter-hateoas>2.6.1</version.spring-boot-starter-hateoas>
         <version.jsoup>1.13.1</version.jsoup>
-        <version.jjwt>0.9.1</version.jjwt>
-        <version.springdoc-openapi-ui>1.6.4</version.springdoc-openapi-ui>
-        <version.commons>1.6.0</version.commons>
-        <org.mapstruct.version>1.4.2.Final</org.mapstruct.version>
+        <version.springdoc-openapi-ui>1.6.8</version.springdoc-openapi-ui>
         <org.projectlombok.version>1.18.22</org.projectlombok.version>
-        <lombok-mapstruct-binding.version>0.2.0</lombok-mapstruct-binding.version>
         <spring-boot-maven-plugin.version>2.4.4</spring-boot-maven-plugin.version>
         <maven-resources-plugin.version>3.2.0</maven-resources-plugin.version>
         <maven-compiler-plugin.version>3.8.1</maven-compiler-plugin.version>
+        <version.keycloak>18.0.0</version.keycloak>
     </properties>
 
     <dependencies>
-        <dependency>
-            <groupId>com.blackphoenixproductions</groupId>
-            <artifactId>commons</artifactId>
-            <version>${version.commons}</version>
-        </dependency>
 
         <dependency>
             <groupId>org.springframework.boot</groupId>
@@ -178,8 +171,8 @@
                 <artifactId>maven-compiler-plugin</artifactId>
                 <version>${maven-compiler-plugin.version}</version>
                 <configuration>
-                    <source>1.8</source> <!-- depending on your project -->
-                    <target>1.8</target> <!-- depending on your project -->
+                    <source>11</source>
+                    <target>11</target>
                     <annotationProcessorPaths>
                         <path>
                             <groupId>org.projectlombok</groupId>
