@@ -1,13 +1,15 @@
 package com.blackphoenixproductions.forumbackend.repository.specification;
 
-import dto.Filter;
-import enums.BooleanOperator;
+import com.blackphoenixproductions.forumbackend.dto.Filter;
+import com.blackphoenixproductions.forumbackend.enums.BooleanOperator;
 import org.springframework.data.jpa.domain.Specification;
 
-import java.text.ParseException;
-import java.text.SimpleDateFormat;
+import javax.persistence.criteria.CriteriaBuilder;
+import javax.persistence.criteria.Predicate;
+import javax.persistence.criteria.Root;
+import java.time.LocalDate;
+import java.time.LocalDateTime;
 import java.util.ArrayList;
-import java.util.Date;
 import java.util.List;
 
 import static org.springframework.data.jpa.domain.Specification.where;
@@ -18,7 +20,8 @@ public class SpecificationBuilder {
         Specification<T> specification = null;
         // figlio
         if (filter.getBooleanOperator() == null
-                && !filter.getValue().isEmpty()) {
+                && ( (filter.getValues() != null && !filter.getValues().isEmpty())
+                || ( filter.getValue() != null && !filter.getValue().isEmpty()) ) ) {
             return createSpecification(filter);
         // padre
         } else if (filter.getBooleanOperator() != null) {
@@ -60,9 +63,50 @@ public class SpecificationBuilder {
                 return (root, query, criteriaBuilder) ->
                         criteriaBuilder.in(root.get(input.getField()))
                                 .value(castToRequiredType(root.get(input.getField()).getJavaType(), input.getValues()));
+            case BETWEEN:
+                return (root, query, criteriaBuilder) ->
+                        getBetweenPredicate(input, root, criteriaBuilder);
+            case EQUALS_TRUNC_DATETIME:
+                return (root, query, criteriaBuilder) -> {
+                    LocalDate paramDate = LocalDate.parse(input.getValue());
+                    LocalDateTime startDate = LocalDateTime.of(paramDate.getYear(), paramDate.getMonth(), paramDate.getDayOfMonth(), 0, 0, 0, 0);
+                    LocalDateTime endDate = LocalDateTime.of(paramDate.getYear(), paramDate.getMonth(), paramDate.getDayOfMonth(), 23, 59, 59, 999999999);
+                    return criteriaBuilder.between(root.get(input.getField()),
+                                startDate,
+                                endDate);
+                };
+
             default:
-                throw new RuntimeException("Operazione non supportata.");
+                throw new RuntimeException("Operazione: " + input.getQueryOperator() + " non supportata.");
         }
+    }
+
+    private <T> Predicate getBetweenPredicate(Filter input, Root<T> root, CriteriaBuilder criteriaBuilder) {
+        Predicate resultPredicate = null;
+        if(root.get(input.getField()).getJavaType().isAssignableFrom(LocalDateTime.class)) {
+            resultPredicate =  criteriaBuilder.between(root.get(input.getField()),
+                    LocalDateTime.parse(input.getValues().get(0)),
+                    LocalDateTime.parse(input.getValues().get(1)));
+        } else if (root.get(input.getField()).getJavaType().isAssignableFrom(LocalDate.class)){
+            resultPredicate =  criteriaBuilder.between(root.get(input.getField()),
+                    LocalDate.parse(input.getValues().get(0)),
+                    LocalDate.parse(input.getValues().get(1)));
+        } else if (root.get(input.getField()).getJavaType().isAssignableFrom(Integer.class)){
+            resultPredicate =  criteriaBuilder.between(root.get(input.getField()),
+                    Integer.valueOf(input.getValues().get(0)),
+                    Integer.valueOf(input.getValues().get(1)));
+        } else if (root.get(input.getField()).getJavaType().isAssignableFrom(Long.class)){
+            resultPredicate =  criteriaBuilder.between(root.get(input.getField()),
+                    Long.valueOf(input.getValues().get(0)),
+                    Long.valueOf(input.getValues().get(1)));
+        } else if (root.get(input.getField()).getJavaType().isAssignableFrom(Double.class)){
+            resultPredicate =  criteriaBuilder.between(root.get(input.getField()),
+                    Double.valueOf(input.getValues().get(0)),
+                    Double.valueOf(input.getValues().get(1)));
+        } else {
+            throw new RuntimeException("Tipo: " + root.get(input.getField()).getJavaType() + " non gestito per operazione BETWEEN.");
+        }
+        return resultPredicate;
     }
 
     private Object castToRequiredType(Class fieldType, String value) {
@@ -70,13 +114,10 @@ public class SpecificationBuilder {
             return Double.valueOf(value);
         }else if(fieldType.isAssignableFrom(Integer.class)){
             return Integer.valueOf(value);
-        }else if(fieldType.isAssignableFrom(Date.class)){
-            try {
-                SimpleDateFormat simpleDateFormat = new SimpleDateFormat("yyyy-MM-dd");
-                return simpleDateFormat.parse(value);
-            } catch (ParseException e){
-                throw new RuntimeException("Formato di data non supportato.");
-            }
+        }else if(fieldType.isAssignableFrom(LocalDate.class)){
+            return LocalDate.parse(value);
+        }else if(fieldType.isAssignableFrom(LocalDateTime.class)){
+            return LocalDateTime.parse(value);
         }
         return value;
     }
