@@ -1,21 +1,18 @@
 package com.blackphoenixproductions.forumbackend.service.impl;
 
+import com.blackphoenixproductions.forumbackend.dto.openApi.exception.CustomException;
+import com.blackphoenixproductions.forumbackend.dto.openApi.post.EditPostDTO;
+import com.blackphoenixproductions.forumbackend.dto.openApi.post.InsertPostDTO;
 import com.blackphoenixproductions.forumbackend.email.EmailSender;
 import com.blackphoenixproductions.forumbackend.entity.Post;
 import com.blackphoenixproductions.forumbackend.entity.Topic;
 import com.blackphoenixproductions.forumbackend.entity.User;
-import com.blackphoenixproductions.forumbackend.mapping.PostMapper;
+import com.blackphoenixproductions.forumbackend.enums.Roles;
 import com.blackphoenixproductions.forumbackend.repository.PostRepository;
 import com.blackphoenixproductions.forumbackend.repository.TopicRepository;
-import com.blackphoenixproductions.forumbackend.repository.UserRepository;
+import com.blackphoenixproductions.forumbackend.security.KeycloakUtility;
 import com.blackphoenixproductions.forumbackend.service.IPostService;
 import com.blackphoenixproductions.forumbackend.service.IUserService;
-import dto.PostDTO;
-import dto.SimpleUserDTO;
-import dto.openApi.exception.CustomException;
-import dto.openApi.post.EditPostDTO;
-import dto.openApi.post.InsertPostDTO;
-import enums.Roles;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -24,7 +21,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import javax.servlet.http.HttpServletRequest;
-import java.text.SimpleDateFormat;
+import java.time.LocalDateTime;
 import java.util.*;
 
 
@@ -32,20 +29,16 @@ import java.util.*;
 public class PostService implements IPostService {
 
     private final PostRepository postRepository;
-    private final UserRepository userRepository;
     private final IUserService userService;
     private final TopicRepository topicRepository;
-    private final PostMapper postMapper;
     private final EmailSender emailSender;
 
 
     @Autowired
-    public PostService(PostRepository postRepository, UserRepository userRepository, IUserService userService, TopicRepository topicRepository, PostMapper postMapper, EmailSender emailSender) {
+    public PostService(PostRepository postRepository, IUserService userService, TopicRepository topicRepository, EmailSender emailSender) {
         this.postRepository = postRepository;
         this.userService = userService;
-        this.postMapper = postMapper;
         this.emailSender = emailSender;
-        this.userRepository = userRepository;
         this.topicRepository = topicRepository;
     }
 
@@ -57,20 +50,17 @@ public class PostService implements IPostService {
 
     @Transactional(readOnly=true)
     @Override
-    public Page<PostDTO> getPagedPosts(Long topicId, Pageable pageable) {
+    public Page<Post> getPagedPosts(Long topicId, Pageable pageable) {
         Topic topic = new Topic(topicId);
-        Page<Post> pagedPosts = postRepository.findByTopicAndDeleteDateIsNull(topic, pageable);
-        Page<PostDTO> pagedPostsDTO = pagedPosts.map(p -> convertPostToPostDTO(p, new SimpleDateFormat("'Inviato il' dd/MM/yy 'alle' HH:mm"),
-                 new SimpleDateFormat("'Ultima modifica il' dd/MM/yy 'alle' HH:mm")));
-        return pagedPostsDTO;
+        return postRepository.findByTopicAndDeleteDateIsNull(topic, pageable);
     }
 
 
     @Transactional
     @Override
-    public PostDTO createPost(InsertPostDTO postDTO) {
-        Post post = postMapper.insertPostDTOtoPost(postDTO);
-        User user = userRepository.findByUsername(postDTO.getUsername());
+    public Post createPost(InsertPostDTO postDTO, HttpServletRequest req) {
+        Post post = new Post();
+        User user = userService.retriveUser(KeycloakUtility.getAccessToken(req));
         if (user == null) {
             throw new CustomException("Utente non trovato.", HttpStatus.NOT_FOUND);
         }
@@ -78,42 +68,33 @@ public class PostService implements IPostService {
         if (!topic.isPresent()){
             throw new CustomException("Topic con id: " + postDTO.getTopicId() + " non trovato.", HttpStatus.NOT_FOUND);
         }
+        post.setMessage(postDTO.getMessage());
         post.setTopic(topic.get());
         post.setUser(user);
-        post.setCreateDate(new Date());
+        post.setCreateDate(LocalDateTime.now());
         Post savedPost = postRepository.save(post);
         // se la risposta non è del creatore del topic e se emailUser = true allora invia un email al creatore del topic
         if (!savedPost.getUser().equals(savedPost.getTopic().getUser()) && savedPost.getTopic().isEmailUser()){
             emailSender.sendTopicReplyEmail(savedPost.getTopic().getUser(), savedPost.getUser(), savedPost) ;
         }
-        return postMapper.postToPostDTO(savedPost);
+        return savedPost;
     }
 
     @Transactional
-    @Override
-    public PostDTO editPost(EditPostDTO postDTO, HttpServletRequest req) {
+    public Post editPost(EditPostDTO postDTO, HttpServletRequest req) {
         Optional<Post> post = postRepository.findById(postDTO.getId());
         if(!post.isPresent()){
             throw new CustomException("Post con id: " + postDTO.getId() + " non trovato.", HttpStatus.BAD_REQUEST);
         }
-        SimpleUserDTO simpleUserDTO = userService.getUserFromToken(req);
-        if (!simpleUserDTO.getRole().equals(Roles.ROLE_STAFF) && !simpleUserDTO.getUsername().equals(post.get().getUser().getUsername())){
+        if(!KeycloakUtility.getRoles(req).contains(Roles.ROLE_STAFF.getValue())
+                && !post.get().getUser().getEmail().equals(KeycloakUtility.getAccessToken(req).getEmail())){
             throw new CustomException("Utente non autorizzato alla modifica del post.", HttpStatus.UNAUTHORIZED);
         }
         Post postToEdit = post.get();
-        postToEdit.setEditDate(new Date());
+        postToEdit.setEditDate(LocalDateTime.now());
         postToEdit.setMessage(postDTO.getMessage());
-        Post updatedPost = postRepository.save(postToEdit);
-        return postMapper.postToPostDTO(updatedPost);
+        return postRepository.save(postToEdit);
     }
 
-    private PostDTO convertPostToPostDTO(Post post, SimpleDateFormat sdf, SimpleDateFormat sdf_edit){
-        PostDTO postDTO = postMapper.postToPostDTO(post);
-        postDTO.setFormattedCreateDate(sdf.format(post.getCreateDate()));
-        if (post.getEditDate() != null){
-            postDTO.setFormattedEditDate(sdf_edit.format(post.getEditDate()));
-        }
-        return postDTO;
-    }
 
 }
