@@ -1,30 +1,22 @@
 package com.blackphoenixproductions.forumbackend.service.impl;
 
+import com.blackphoenixproductions.forumbackend.dto.openApi.exception.CustomException;
+import com.blackphoenixproductions.forumbackend.dto.openApi.topic.EditTopicDTO;
+import com.blackphoenixproductions.forumbackend.dto.openApi.topic.InsertTopicDTO;
 import com.blackphoenixproductions.forumbackend.entity.Topic;
 import com.blackphoenixproductions.forumbackend.entity.User;
-import com.blackphoenixproductions.forumbackend.mapping.TopicMapper;
+import com.blackphoenixproductions.forumbackend.enums.Roles;
 import com.blackphoenixproductions.forumbackend.repository.TopicRepository;
-import com.blackphoenixproductions.forumbackend.repository.UserRepository;
+import com.blackphoenixproductions.forumbackend.security.KeycloakUtility;
 import com.blackphoenixproductions.forumbackend.service.ITopicService;
 import com.blackphoenixproductions.forumbackend.service.IUserService;
-import com.blackphoenixproductions.forumbackend.utility.DateUtility;
-import dto.SimpleTopicDTO;
-import dto.SimpleUserDTO;
-import dto.TopicDTO;
-import dto.openApi.exception.CustomException;
-import dto.openApi.topic.EditTopicDTO;
-import dto.openApi.topic.InsertTopicDTO;
-import enums.Roles;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import javax.servlet.http.HttpServletRequest;
-import java.text.SimpleDateFormat;
-import java.util.Date;
+import java.time.LocalDateTime;
 import java.util.Optional;
 
 
@@ -33,16 +25,12 @@ import java.util.Optional;
 public class TopicService implements ITopicService {
 
     private final TopicRepository topicRepository;
-    private final UserRepository userRepository;
     private final IUserService userService;
-    private final TopicMapper topicMapper;
 
     @Autowired
-    public TopicService(TopicRepository topicRepository, UserRepository userRepository, IUserService userService, TopicMapper topicMapper) {
+    public TopicService(TopicRepository topicRepository, IUserService userService) {
         this.topicRepository = topicRepository;
-        this.userRepository = userRepository;
         this.userService = userService;
-        this.topicMapper = topicMapper;
     }
 
     @Transactional(readOnly=true)
@@ -52,38 +40,27 @@ public class TopicService implements ITopicService {
     }
 
 
-    @Transactional(readOnly=true)
-    @Override
-    public Page<SimpleTopicDTO> getPagedTopics(Pageable pageable, String title, String username) {
-        Page<Topic> pagedTopics = null;
-        if (username != null && !username.isEmpty()){
-            User findedUser = userRepository.findByUsername(username);
-            pagedTopics = topicRepository.findByDeleteDateIsNullAndTitleIgnoreCaseContainingAndUser(pageable, title, findedUser);
-        } else {
-            pagedTopics = topicRepository.findByDeleteDateIsNullAndTitleIgnoreCaseContaining(pageable, title);
-        }
-        return pagedTopics.map(p -> convertTopicToSimpleTopicDTO(p));
-    }
-
     @Transactional
     @Override
-    public SimpleTopicDTO createTopic(InsertTopicDTO insertTopicDTO) {
-        User userTopic = userRepository.findByUsername(insertTopicDTO.getUsername());
+    public Topic createTopic(InsertTopicDTO insertTopicDTO, HttpServletRequest req) {
+        User userTopic = userService.retriveUser(KeycloakUtility.getAccessToken(req));
         if (userTopic == null){
             throw new CustomException("Utente non trovato.", HttpStatus.NOT_FOUND);
         }
-        Topic topic = topicMapper.insertTopicDTOtoTopic(insertTopicDTO);
+        Topic topic = new Topic();
+        topic.setTitle(insertTopicDTO.getTitle());
+        topic.setMessage(insertTopicDTO.getMessage());
+        topic.setEmailUser(insertTopicDTO.isEmailUser());
+        topic.setPinned(insertTopicDTO.isPinned());
         topic.setUser(userTopic);
-        topic.setCreateDate(new Date());
-        Topic savedTopic = topicRepository.save(topic);
-        return topicMapper.topicToSimpleTopicDTO(savedTopic);
+        topic.setCreateDate(LocalDateTime.now());
+        return topicRepository.save(topic);
     }
 
     @Transactional
     @Override
-    public TopicDTO editTopic(EditTopicDTO topicDTO, HttpServletRequest req) {
-        SimpleUserDTO simpleUserDTO = userService.getUserFromToken(req);
-        if (!simpleUserDTO.getRole().equals(Roles.ROLE_STAFF)){
+    public Topic editTopic(EditTopicDTO topicDTO, HttpServletRequest req) {
+        if(!KeycloakUtility.getRoles(req).contains(Roles.ROLE_STAFF.getValue())){
             throw new CustomException("Utente non autorizzato alla modifica di un topic.", HttpStatus.UNAUTHORIZED);
         }
         Optional<Topic> topic = topicRepository.findById(topicDTO.getId());
@@ -91,43 +68,19 @@ public class TopicService implements ITopicService {
             throw new CustomException("Topic con id: " + topicDTO.getId() + " non trovato.", HttpStatus.BAD_REQUEST);
         }
         Topic topicToEdit = topic.get();
-        topicToEdit.setEditDate(new Date());
+        topicToEdit.setEditDate(LocalDateTime.now());
         topicToEdit.setMessage(topicDTO.getMessage());
-        Topic updatedTopic = topicRepository.save(topicToEdit);
-        return topicMapper.topicToTopicDTO(updatedTopic);
+        return topicRepository.save(topicToEdit);
     }
 
     @Transactional(readOnly=true)
     @Override
-    public TopicDTO getTopic(Long id) {
+    public Topic getTopic(Long id) {
         Optional<Topic> topic = topicRepository.findById(id);
         if(!topic.isPresent()){
             throw new CustomException("Topic con id: " + id + " non trovato.", HttpStatus.BAD_REQUEST);
         }
-        TopicDTO topicDTO = topicMapper.topicToTopicDTO(topic.get());
-        setFormattedDate(topicDTO, topic.get());
-        if (topic.get().getEditDate() != null) {
-            setFormattedEditDate(topicDTO, topic.get());
-        }
-        return topicDTO;
-    }
-
-    private void setFormattedDate(TopicDTO topicDTO, Topic savedTopic) {
-        SimpleDateFormat sdf = new SimpleDateFormat("'Inviato il' dd/MM/yy 'alle' HH:mm");
-        topicDTO.setFormattedDate(sdf.format(savedTopic.getCreateDate()));
-    }
-
-    private void setFormattedEditDate(TopicDTO topicDTO, Topic savedTopic) {
-        SimpleDateFormat sdf = new SimpleDateFormat("'Ultima modifica il' dd/MM/yy 'alle' HH:mm");
-        topicDTO.setFormattedEditDate(sdf.format(savedTopic.getEditDate()));
-    }
-
-
-    private SimpleTopicDTO convertTopicToSimpleTopicDTO(Topic topic){
-        SimpleTopicDTO simpleTopicDTO = topicMapper.topicToSimpleTopicDTO(topic);
-        simpleTopicDTO.setPostsNumber(topic.getPosts().size());
-        simpleTopicDTO.setTimeDifferenceFromNow(DateUtility.setTimeDifferenceFromNow(simpleTopicDTO.getCreateDate()));
-        return simpleTopicDTO;
+        return topic.get();
     }
 
 
