@@ -1,28 +1,25 @@
 package com.blackphoenixproductions.forumbackend.api;
 
 
+import com.blackphoenixproductions.forumbackend.dto.NotificationDTO;
+import com.blackphoenixproductions.forumbackend.entity.User;
+import com.blackphoenixproductions.forumbackend.security.KeycloakUtility;
 import com.blackphoenixproductions.forumbackend.service.INotificationService;
 import com.blackphoenixproductions.forumbackend.service.IUserService;
-import dto.NotificationDTO;
-import dto.SimpleUserDTO;
 import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
-import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.hateoas.CollectionModel;
-import org.springframework.hateoas.EntityModel;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpServletRequest;
 import java.util.List;
@@ -50,22 +47,19 @@ public class NotificationRestAPIController {
     @ApiResponses(value = {
             @ApiResponse(responseCode = "200", description = "OK"),
             @ApiResponse(responseCode = "403", description = "Forbidden.", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "401", description = "Unauthorized: JWT token scaduto oppure non valido.", content = @Content(schema = @Schema(hidden = true)))
+            @ApiResponse(responseCode = "401", description = "Unauthorized", content = @Content(schema = @Schema(hidden = true)))
     })
-    @Operation(summary = "Restituisce tutte le notifiche di un utente.", security = @SecurityRequirement(name = "bearerAuth"))
+    @Operation(summary = "Restituisce tutte le notifiche di un utente.")
     @GetMapping(value = "getUserNotificationList")
-    @PreAuthorize("hasRole('ROLE_STAFF') or hasRole('ROLE_USER') or hasRole('ROLE_FACEBOOK') or hasRole('ROLE_GOOGLE')")
     public ResponseEntity<CollectionModel<NotificationDTO>> getUserNotificationList(HttpServletRequest req){
         logger.info("Start getUserNotificationList");
         CollectionModel<NotificationDTO> userNotificationModel = null;
-        SimpleUserDTO simpleUserDTO = userService.getUserFromToken(req);
-        List<NotificationDTO> userNotification = notificationService.getUserNotification(simpleUserDTO);
-//        List<EntityModel<NotificationDTO>> userNotificationModel = userNotification.stream().
-//                map(n -> EntityModel.of(n, linkTo(methodOn(NotificationRestAPIController.class).getUserNotificationList(req)).withSelfRel()))
-//                .collect(Collectors.toList());
-//        CollectionModel<EntityModel<NotificationDTO>> response = CollectionModel.of(userNotification, linkTo(methodOn(NotificationRestAPIController.class).getUserNotificationList(req)).withSelfRel());
-        if(userNotification != null) {
-            userNotificationModel = CollectionModel.of(userNotification, linkTo(methodOn(NotificationRestAPIController.class).getUserNotificationList(req)).withSelfRel());
+        User user = userService.getUserFromEmail(KeycloakUtility.getAccessToken(req).getEmail());
+        if(user != null) {
+            List<NotificationDTO> userNotification = notificationService.getUserNotification(user);
+            if (userNotification != null) {
+                userNotificationModel = CollectionModel.of(userNotification, linkTo(methodOn(NotificationRestAPIController.class).getUserNotificationList(req)).withSelfRel());
+            }
         }
         logger.info("End getUserNotificationList");
         return new ResponseEntity<CollectionModel<NotificationDTO>>(userNotificationModel, HttpStatus.OK);
@@ -75,13 +69,15 @@ public class NotificationRestAPIController {
             @ApiResponse(responseCode = "200", description = "OK"),
             @ApiResponse(responseCode = "403", description = "Forbidden.", content = @Content(schema = @Schema(hidden = true))),
     })
-    @Operation(summary = "Restituisce lo status delle notifiche di un utente. Permette di capire se un utente ha almeno una notfica da leggere.", security = @SecurityRequirement(name = "bearerAuth"))
+    @Operation(summary = "Restituisce lo status delle notifiche di un utente. Permette di capire se un utente ha almeno una notfica da leggere.")
     @GetMapping(value = "getUserNotificationStatus")
-    @PreAuthorize("hasRole('ROLE_STAFF') or hasRole('ROLE_USER') or hasRole('ROLE_FACEBOOK') or hasRole('ROLE_GOOGLE')")
     public ResponseEntity<Boolean> getUserNotificationStatus(HttpServletRequest req){
         logger.info("Start getUserNotificationStatus");
-        SimpleUserDTO simpleUserDTO = userService.getUserFromToken(req);
-        Boolean notificationStatus = notificationService.getUserNotificationStatus(simpleUserDTO);
+        Boolean notificationStatus = null;
+        User user = userService.getUserFromEmail(KeycloakUtility.getAccessToken(req).getEmail());
+        if(user != null) {
+            notificationStatus = notificationService.getUserNotificationStatus(user);
+        }
         logger.info("End getUserNotificationStatus");
         return new ResponseEntity<Boolean>(notificationStatus, HttpStatus.OK);
     }
@@ -90,14 +86,15 @@ public class NotificationRestAPIController {
             @ApiResponse(responseCode = "200", description = "OK"),
             @ApiResponse(responseCode = "403", description = "Forbidden.", content = @Content(schema = @Schema(hidden = true))),
     })
-    @Operation(summary = "Imposta lo status delle notifiche di un utente a 'tutte lette'.", security = @SecurityRequirement(name = "bearerAuth"))
-    @GetMapping(value = "setReadedNotificationStatus")
-    @PreAuthorize("hasRole('ROLE_STAFF') or hasRole('ROLE_USER') or hasRole('ROLE_FACEBOOK') or hasRole('ROLE_GOOGLE')")
-    public ResponseEntity<String> setReadedNotificationStatus(HttpServletRequest req){
-        logger.info("Start setReadedNotificationStatus");
-        SimpleUserDTO simpleUserDTO = userService.getUserFromToken(req);
-        notificationService.setReadedNotificationStatus(simpleUserDTO);
-        logger.info("End setReadedNotificationStatus");
+    @Operation(summary = "Imposta lo status delle notifiche di un utente.")
+    @PostMapping(value = "setNotificationStatus")
+    public ResponseEntity<String> setNotificationStatus(HttpServletRequest req, @Parameter(description = "Determina se mostrare l'avviso di una nuova notifica da leggere.") @RequestParam boolean showNotificationNotice){
+        logger.info("Start setNotificationStatus");
+        User user = userService.getUserFromEmail(KeycloakUtility.getAccessToken(req).getEmail());
+        if(user != null) {
+            notificationService.setNotificationStatus(user.getUsername(), showNotificationNotice);
+        }
+        logger.info("End setNotificationStatus");
         return new ResponseEntity<String>("Le notifiche sono state lette.", HttpStatus.OK);
     }
 
