@@ -1,10 +1,6 @@
 package com.blackphoenixproductions.forumbackend.api;
 
 import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.media.Content;
-import io.swagger.v3.oas.annotations.media.Schema;
-import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
@@ -15,6 +11,8 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import javax.servlet.http.HttpServletRequest;
+
+
 
 @RestController
 @RequestMapping("/api")
@@ -29,7 +27,7 @@ public class ValueRestAPIController {
     }
 
 
-    @Operation(summary = "Restituisce la versione del forum-backend.")
+    @Operation(summary = "Restituisce la versione del forum-backend.", hidden = true)
     @GetMapping(value = "/getBuildVersionBackEnd")
     public ResponseEntity<String> getBuildVersionBackEnd (HttpServletRequest req){
         return new ResponseEntity<String>(buildVersion, HttpStatus.OK);
