@@ -1,8 +1,8 @@
 package com.blackphoenixproductions.forumbackend.assembler;
 
 import com.blackphoenixproductions.forumbackend.api.PostRestAPIController;
-import dto.PostDTO;
-import dto.openApi.post.EditPostDTO;
+import com.blackphoenixproductions.forumbackend.dto.openApi.post.EditPostDTO;
+import com.blackphoenixproductions.forumbackend.entity.Post;
 import org.springframework.hateoas.EntityModel;
 import org.springframework.hateoas.server.RepresentationModelAssembler;
 import org.springframework.stereotype.Component;
@@ -11,9 +11,9 @@ import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.linkTo;
 import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.methodOn;
 
 @Component
-public class PostDTOAssembler implements RepresentationModelAssembler<PostDTO, EntityModel<PostDTO>> {
+public class PostAssembler implements RepresentationModelAssembler<Post, EntityModel<Post>> {
     @Override
-    public EntityModel<PostDTO> toModel(PostDTO entity) {
+    public EntityModel<Post> toModel(Post entity) {
         return EntityModel.of(entity, linkTo(methodOn(PostRestAPIController.class).editPost(new EditPostDTO(entity.getId()), null)).withSelfRel());
     }
 }
