@@ -1,7 +1,7 @@
 package com.blackphoenixproductions.forumbackend.assembler;
 
 import com.blackphoenixproductions.forumbackend.api.TopicRestAPIController;
-import dto.SimpleTopicDTO;
+import com.blackphoenixproductions.forumbackend.entity.Topic;
 import org.springframework.hateoas.EntityModel;
 import org.springframework.hateoas.server.RepresentationModelAssembler;
 import org.springframework.stereotype.Component;
@@ -10,9 +10,9 @@ import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.linkTo;
 import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.methodOn;
 
 @Component
-public class SimpleTopicDTOAssembler implements RepresentationModelAssembler<SimpleTopicDTO, EntityModel<SimpleTopicDTO>> {
+public class TopicAssembler implements RepresentationModelAssembler<Topic, EntityModel<Topic>> {
     @Override
-    public EntityModel<SimpleTopicDTO> toModel(SimpleTopicDTO entity) {
+    public EntityModel<Topic> toModel(Topic entity) {
         return EntityModel.of(entity, linkTo(methodOn(TopicRestAPIController.class).findTopic(entity.getId())).withSelfRel());
 //                linkTo(methodOn(TopicRestAPIController.class).findTopic(entity.getId())).withRel(""));
     }
