@@ -1,8 +1,6 @@
 package com.blackphoenixproductions.forumbackend.repository.projection;
 
 public interface IUser {
-
     String getUsername();
-    String getRole();
-
+    String getEmail();
 }
