@@ -1,19 +1,11 @@
 package com.blackphoenixproductions.forumbackend.security;
 
 import io.swagger.v3.oas.annotations.OpenAPIDefinition;
-import io.swagger.v3.oas.annotations.enums.SecuritySchemeType;
+import io.swagger.v3.oas.annotations.info.Contact;
 import io.swagger.v3.oas.annotations.info.Info;
-import io.swagger.v3.oas.annotations.security.SecurityScheme;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
-@OpenAPIDefinition(info = @Info(title = "Whiteboard API", version = "v1", description = "N.B.: Per utilizzare gli endpoint protetti è prima necessario chiamare l'endpoint /login dopodiché incollare l'accessToken ottenuto nella finestra di Authorize."))
-@SecurityScheme(
-        name = "bearerAuth",
-        type = SecuritySchemeType.HTTP,
-        bearerFormat = "JWT",
-        scheme = "bearer"
-)
+@OpenAPIDefinition(info = @Info(title = "Forum API", version = "v1", description = "Le API del forum.", contact = @Contact(email = "zampetti1@hotmail.com")))
 public class OpenApi30Config {
-
 }
