@@ -1,12 +1,8 @@
 package com.blackphoenixproductions.forumbackend.service;
 
-
-import dto.SimpleTopicDTO;
-import dto.TopicDTO;
-import dto.openApi.topic.EditTopicDTO;
-import dto.openApi.topic.InsertTopicDTO;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
+import com.blackphoenixproductions.forumbackend.dto.openApi.topic.EditTopicDTO;
+import com.blackphoenixproductions.forumbackend.dto.openApi.topic.InsertTopicDTO;
+import com.blackphoenixproductions.forumbackend.entity.Topic;
 
 import javax.servlet.http.HttpServletRequest;
 
@@ -14,11 +10,9 @@ public interface ITopicService {
 
     Long getTotalTopics();
 
-    Page<SimpleTopicDTO> getPagedTopics(Pageable pageable, String title, String username);
+    Topic createTopic(InsertTopicDTO insertTopicDTO, HttpServletRequest req);
 
-    SimpleTopicDTO createTopic(InsertTopicDTO insertTopicDTO);
+    Topic editTopic(EditTopicDTO topicDTO, HttpServletRequest req);
 
-    TopicDTO editTopic(EditTopicDTO topicDTO, HttpServletRequest req);
-
-    TopicDTO getTopic (Long id);
+    Topic getTopic(Long id);
 }
