@@ -1,7 +1,7 @@
 package com.blackphoenixproductions.forumbackend.service;
 
+import com.blackphoenixproductions.forumbackend.dto.Filter;
 import com.blackphoenixproductions.forumbackend.entity.VTopic;
-import dto.Filter;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
