@@ -1,141 +1,122 @@
 package com.blackphoenixproductions.forumbackend.service.impl;
 
-import com.blackphoenixproductions.forumbackend.mapping.UserMapper;
+import com.blackphoenixproductions.forumbackend.dto.openApi.exception.CustomException;
+import com.blackphoenixproductions.forumbackend.enums.Roles;
 import com.blackphoenixproductions.forumbackend.repository.UserRepository;
+import com.blackphoenixproductions.forumbackend.security.KeycloakUtility;
 import com.blackphoenixproductions.forumbackend.service.IUserService;
-import dto.SimpleUserDTO;
-import dto.UserDTO;
 import com.blackphoenixproductions.forumbackend.entity.User;
-import com.blackphoenixproductions.forumbackend.security.JwtTokenProvider;
-import dto.openApi.exception.CustomException;
-import enums.Roles;
+import org.keycloak.admin.client.Keycloak;
+import org.keycloak.admin.client.KeycloakBuilder;
+import org.keycloak.representations.AccessToken;
+import org.keycloak.representations.idm.UserRepresentation;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.http.*;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import javax.servlet.http.HttpServletRequest;
+import java.util.Set;
+
 
 @Service
 public class UserService implements IUserService {
 
-    private final UserMapper userMapper;
+    private static final Logger logger = LoggerFactory.getLogger(UserService.class);
+
     private final UserRepository userRepository;
-    private final PasswordEncoder passwordEncoder;
-    private final JwtTokenProvider jwtTokenProvider;
-    private final AuthenticationManager authenticationManager;
+    private String KEYCLOAK_SERVER_URL;
+    private String KEYCLOAK_REALM;
+    private String KEYCLOAK_RESOURCE;
+    private String SERVICE_USER_USERNAME;
+    private String SERVICE_USER_PASSWORD;
 
     @Autowired
-    public UserService(UserMapper userMapper, UserRepository userRepository, PasswordEncoder passwordEncoder, JwtTokenProvider jwtTokenProvider, AuthenticationManager authenticationManager) {
-        this.userMapper = userMapper;
+    public UserService(UserRepository userRepository,
+                       @Value("${keycloak.auth-server-url}") String KEYCLOAK_SERVER_URL,
+                       @Value("${keycloak.realm}") String KEYCLOAK_REALM,
+                       @Value("${keycloak.resource}") String KEYCLOAK_RESOURCE,
+                       @Value("${service.user.username}") String SERVICE_USER_USERNAME,
+                       @Value("${service.user.password}") String SERVICE_USER_PASSWORD)  {
         this.userRepository = userRepository;
-        this.passwordEncoder = passwordEncoder;
-        this.jwtTokenProvider = jwtTokenProvider;
-        this.authenticationManager = authenticationManager;
+        this.KEYCLOAK_SERVER_URL = KEYCLOAK_SERVER_URL;
+        this.KEYCLOAK_REALM = KEYCLOAK_REALM;
+        this.KEYCLOAK_RESOURCE = KEYCLOAK_RESOURCE;
+        this.SERVICE_USER_USERNAME = SERVICE_USER_USERNAME;
+        this.SERVICE_USER_PASSWORD = SERVICE_USER_PASSWORD;
     }
 
+    /**
+     * Recupera l'utente registrato.
+     * Nota bene: l'utente viene persistito nell'applicativo durante la registrazione con Keycloak tramite un plugin custom,
+     * tuttavia se non dovesse trovarlo (anomalia o demo) lo salva ora.
+     * @param accessToken
+     * @return
+     */
     @Override
     @Transactional
-    public String login(String username, String password) {
-        try {
-            authenticationManager.authenticate(new UsernamePasswordAuthenticationToken(username, password));
-            User findedUser = userRepository.findByUsername(username);
-            return jwtTokenProvider.createToken(username, findedUser.getRole());
-        } catch (AuthenticationException e) {
-            throw new CustomException("Username/password non valido", HttpStatus.UNPROCESSABLE_ENTITY);
-        }
-    }
-
-    @Override
-    @Transactional
-    public String signin(UserDTO userDTO) {
-        User findedUser = userRepository.findByUsernameOrEmail(userDTO.getUsername(), userDTO.getEmail());
+    public User retriveUser(AccessToken accessToken) {
+        User findedUser = userRepository.findByEmail(accessToken.getEmail());
         if(findedUser == null) {
-            User user = userMapper.userDTOtoUser(userDTO);
-            user.setPassword(passwordEncoder.encode(user.getPassword()));
-            userRepository.save(user);
-            return jwtTokenProvider.createToken(user.getUsername(), user.getRole());
-        } else {
-            throw new CustomException("Username e/o Email gia' utilizzati", HttpStatus.CONFLICT);
+            logger.warn("Utente non trovato nell'applicativo, procedo a salvarlo...");
+            findedUser = userRepository.saveAndFlush(new User(accessToken.getPreferredUsername(), accessToken.getEmail(), getApplicationRole(KeycloakUtility.getRoles(accessToken))));
         }
+        return findedUser;
+    }
+
+    private String getApplicationRole(Set<String> roles) {
+        if(roles.contains(Roles.ROLE_HELPDESK.getValue())){
+            return Roles.ROLE_HELPDESK.getValue();
+        } else if(roles.contains(Roles.ROLE_STAFF.getValue())){
+            return Roles.ROLE_STAFF.getValue();
+        }
+        return Roles.ROLE_USER.getValue();
     }
 
     @Override
     @Transactional(readOnly = true)
-    public SimpleUserDTO getUserFromToken(HttpServletRequest req) {
-        // ricavo username dal token
-        return userRepository.findUserByUsername(jwtTokenProvider.getUsername(jwtTokenProvider.resolveToken(req)));
+    public User getUserFromUsername(String username) {
+        return userRepository.findByUsername(username);
     }
 
     @Override
     @Transactional(readOnly = true)
-    public SimpleUserDTO getUserFromUsername(String username) {
-        return userRepository.findUserByUsername(username);
+    public User getUserFromEmail(String email) {
+        return userRepository.findByEmail(email);
     }
 
     @Override
-    @Transactional(readOnly = true)
-    public UserDTO getUserFromEmail(String email) {
-        return userRepository.findUserByEmail(email);
+    @Transactional
+    public User changeUserUsername(AccessToken accessToken, String username) {
+        User user = retriveUser(accessToken);
+        user.setUsername(username);
+        user = userRepository.saveAndFlush(user);
+        try {
+            Keycloak kc = KeycloakBuilder.builder()
+                    .serverUrl(KEYCLOAK_SERVER_URL)
+                    .realm(KEYCLOAK_REALM)
+                    .username(SERVICE_USER_USERNAME)
+                    .password(SERVICE_USER_PASSWORD)
+                    .clientId(KEYCLOAK_RESOURCE)
+                    .build();
+
+            UserRepresentation keycloakUser = kc.realm(KEYCLOAK_REALM).users().get(accessToken.getSubject()).toRepresentation();
+            keycloakUser.setUsername(username);
+            kc.realm(KEYCLOAK_REALM).users().get(accessToken.getSubject()).update(keycloakUser);
+
+        } catch (Exception e) {
+            throw new CustomException(e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+        return user;
     }
 
-    @Override
-    public String getJwtToken(String username, String role) {
-        return jwtTokenProvider.createToken(username, role);
-    }
 
     @Override
     @Transactional(readOnly = true)
     public Long getTotalUsers() {
         return userRepository.count();
-    }
-
-    @Override
-    public String refresh(String username) {
-        return jwtTokenProvider.createRefreshToken(username);
-    }
-
-    @Override
-    public String getResetToken(String username, String password) {
-        return jwtTokenProvider.createResetToken(username, password );
-    }
-
-    @Override
-    @Transactional(readOnly = true)
-    public String getNewAccessTokenFromRefreshToken(HttpServletRequest req) {
-        String newAccessToken = null;
-        SimpleUserDTO simpleUserDTO = getUserFromToken(req);
-        if (simpleUserDTO != null)
-            newAccessToken = jwtTokenProvider.createToken(simpleUserDTO.getUsername(), simpleUserDTO.getRole());
-        return newAccessToken;
-    }
-
-    @Override
-    @Transactional
-    public User finishResetCredentials(String password, User user) {
-        user.setPassword(passwordEncoder.encode(password));
-        return userRepository.save(user);
-    }
-
-    @Override
-    public boolean isValidResetToken(String token, User user) {
-        return jwtTokenProvider.validateResetToken(token, user.getPassword());
-    }
-
-    @Override
-    @Transactional(readOnly = true)
-    public User getUserFromEmailWithRoleUser(String email) {
-        return userRepository.findByEmailAndRole(email, Roles.ROLE_USER.getValue());
-    }
-
-    @Override
-    @Transactional(readOnly = true)
-    public User getUserFromUsernameWithRoleUser(String username) {
-        return userRepository.findByEmailAndRole(username, Roles.ROLE_USER.getValue());
     }
 
 
